Moved persistance into initializers
