+++ src/js/initializers/shortcuts.js
@@ -1,7 +1,7 @@
 import {ipcRenderer} from "electron"
 import {pinSearchBar, removeAllSearchBarPins} from "../actions/searchBar"
 import {toggleLeftSidebar, toggleRightSidebar} from "../actions/view"
-import {clearState} from "../persistance"
+import {clearState} from "./persistance"
 
 export default store => {
   ipcRenderer.on("pinSearch", () => store.dispatch(pinSearchBar()))
