+++ src/js/initializers/store.js
@@ -4,7 +4,7 @@ import {createStore, applyMiddleware} from "redux"
 import {composeWithDevTools} from "redux-devtools-extension"
 import reducer from "../reducers"
 import reduxThunk from "redux-thunk"
-import {loadState, saveState} from "../persistance"
+import {loadState, saveState} from "./persistance"
 import throttle from "lodash/throttle"
 import Client from "boom-js-client"
 import {getCredentials} from "../reducers/boomdCredentials"
